requirements.txt
flask==3.0.0
pyyaml==6.0.1
python-hcl2==4.3.2
python-dotenv==1.0.0
reportlab==4.0.9
requests==2.31.0
gunicorn==21.2.0